+++ enpt/processors/atmospheric_correction/atmospheric_correction.py
@@ -190,6 +190,13 @@ class AtmosphericCorrector(object):
                 polymer_ac_enmap(enmap_l1b=enmap_ImageL1,
                                  config=self.cfg,
                                  detector='vnir')
+
+            # Overwrite SWIR with 0 for water pixels (POLYMER does not produce a SWIR output)
+            # and NaNs for all other pixels (NaNs are later set to no-data)
+            # -> not needed anymore if implemented in ACwater - https://gitlab.awi.de/phytooptics/acwater/-/issues/23
+            wl_ref_swir = np.zeros_like(wl_ref_swir)
+            wl_ref_swir[enmap_ImageL1.swir.mask_landwater[:] != 2] = np.nan
+
         except:  # noqa
             enmap_ImageL1.logger.error(
                 "The atmospheric correction for water surfaces based on ACwater/Polymer failed (issue tracker at "
@@ -237,6 +244,13 @@ class AtmosphericCorrector(object):
                 polymer_ac_enmap(enmap_l1b=enmap_ImageL1,
                                  config=self.cfg,
                                  detector='vnir')
+
+            # Overwrite SWIR with 0 for water pixels (POLYMER does not produce a SWIR output)
+            # and NaNs for all other pixels (NaNs are later set to no-data)
+            # -> not needed anymore if implemented in ACwater - https://gitlab.awi.de/phytooptics/acwater/-/issues/23
+            wl_ref_swir_water = np.zeros_like(wl_ref_swir_water)
+            wl_ref_swir_water[enmap_ImageL1.swir.mask_landwater[:] != 2] = np.nan
+
         except:  # noqa
             enmap_ImageL1.logger.error(
                 "The atmospheric correction for water surfaces based on ACwater/Polymer failed (issue tracker at "